Added a "favorite weapon" stat
Added a "favorite weapon" stat that will display alongside the normal stats. This stat gets the name of the weapon a player has the most kills with and displays it to them along with the total number of kills they have with the weapon. Eg. "Favorite weapon: MP40 at 1350 kills." Most guns in the game are included along with grenades and the most-common high-kill tank/AT guns. More damage type entries will be added as they are needed but 99% of players' top weapons should already be included.

Also removed the friendly fire stats to help fit the bot message on 1 line at the request of dirtybirdy. A friendly fire stat displaying the ratio between normal kills and friendly fire kills will be added soon.